All of the operations are more consistent now - had overlooked the fact that ensime only works on the file on disk so the file have to be saved first.
Motivation:
All the ensime operations work on the file on disk. Because of this textmate
needs to save the file to disk before calling any operations to make sure
that the most recent changes are used. However, this feature is a bit buggy
and textmate will complain that another process (ENSIME) changed the file
while there were still unsaved edits in the buffer (even though there weren't).
So it isn't possible to save the document, run the ensime command and tell
textmate to reload it without getting an error so another approach was needed

Solution:
The flow is now like this: It will save the current buffer to the HDD. Send
a request to ENSIME to do some operation on the file on disk. ENSIME edits
the file on disk and TextMate reads the new contens of the file and uses that
in the buffer instead.